add dual encoder to the pytorch layers
there isnt yet a TF impl, but this makes it
at least reusable for applications
